requirements.txt
PyQT5
PyYAML
ffprobe-python
NodeGraphQt